requirements.txt
pandas
mysql-connector-python
gradio